+++ datastore/test/utils/test_utils.dart
@@ -1,14 +1,40 @@
+import 'package:objectid/objectid.dart';
 import 'package:oobium_datastore/src/adapters.dart';
 import 'package:oobium_datastore/src/datastore.dart';
 import 'package:oobium_datastore/src/datastore/data.dart';
 import 'package:oobium_datastore/src/datastore/models.dart';
 import 'package:oobium_datastore/src/datastore/repo.dart';
 
-import 'test_models.dart';
-
 final datastores = <DataStore>[];
 final data = <Data>[];
 final repo = <Repo>[];
+
+class TestType1 extends DataModel {
+  ObjectId get id => this['_modelId'];
+  String? get name => this['name'];
+  TestType1({String? name}) : super({'name': name});
+  TestType1._(map) : super(map);
+  TestType1.copyNew(TestType1 original, {String? name}) : super.copyNew(original, {'name': name});
+  TestType1.copyWith(TestType1 original, {String? name}) : super.copyWith(original, {'name': name});
+  TestType1.deleted(TestType1 original) : super.deleted(original);
+  TestType1 copyNew({String? name}) => TestType1.copyNew(this, name: name);
+  TestType1 copyWith({String? name}) => TestType1.copyWith(this, name: name);
+  TestType1 deleted() => TestType1.deleted(this);
+}
+
+class TestType2 extends DataModel {
+  ObjectId get id => this['_modelId'];
+  String? get name => this['name'];
+  TestType1? get type1 => this['type1'];
+  TestType2({String? name, TestType1? type1}) : super({'name': name});
+  TestType2._(map) : super(map);
+  TestType2.copyNew(TestType2 original, {String? name, TestType1? type1}) : super.copyNew(original, {'name': name, 'type1': type1});
+  TestType2.copyWith(TestType2 original, {String? name, TestType1? type1}) : super.copyWith(original, {'name': name, 'type1': type1});
+  TestType2.deleted(TestType2 original) : super.deleted(original);
+  TestType2 copyNew({String? name, TestType1? type1}) => TestType2.copyNew(this, name: name, type1: type1);
+  TestType2 copyWith({String? name, TestType1? type1}) => TestType2.copyWith(this, name: name, type1: type1);
+  TestType2 deleted() => TestType2.deleted(this);
+}
 
 DataStore createDatastore(String testFile, {
   DataStore? clone,
@@ -22,12 +48,12 @@ DataStore createDatastore(String testFile, {
     isolate: isolate,
     adapters: Adapters([
       Adapter<TestType1>(
-        decode: (map) => TestType1.fromJson(map),
+        decode: (map) => TestType1._(map),
         encode: (k,v) => v,
         fields: ['name']
       ),
       Adapter<TestType2>(
-        decode: (map) => TestType2.fromJson(map),
+        decode: (map) => TestType2._(map),
         encode: (k,v) => v,
         fields: ['name','type1']
       )