+++ datastore/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0-dev.2
+
+* add model.deleted method and private constructor
+* convert getAll to return a List
+* make DataModel abstract
+* updated deleted functionality and make .deleted() method abstract
+
 ## 1.0.0-dev.1
 
 * Initial Open Source release.