+++ datastore_gen/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0-dev.2
+
+* updates for oobium_datastore 1.0.0-dev.2
+
 ## 1.0.0-dev.1
 
 * full generation of schema classes