+++ routing_gen/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0-dev.2
+
+* make RouteParser and RouterDelegate final fields of the generated Routes class
+
 ## 1.0.0-dev.1
 
 * full generation of route classes, guaranteeing compile-time safety of urls and deep links